# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mawaeedak API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /themes/{id}:
    patch:
      operationId: updateTheme
      tags: [themes]
      summary: Update theme (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThemeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Theme"

  # News
...
  /admin/social/settings:
    get:
      operationId: getSocialSettings
      tags: [social]
      summary: Get X/Twitter automation settings (admin)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialSettings"
    patch:
      operationId: updateSocialSettings
      tags: [social]
      summary: Update X/Twitter automation settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialSettings"
...
components:
  schemas:
...
    Theme:
      type: object
      required: [id, name, slug, is_active, is_available, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        colors: { type: object }
        is_active: { type: boolean }
        is_available: { type: boolean }
        tier: { type: string }
        created_at: { type: string }
    ThemeUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        colors: { type: object }
        is_active: { type: boolean }
        is_available: { type: boolean }
        tier: { type: string }
...
    SocialSettings:
      type: object
      required: [id, platform, is_enabled, post_time, template, updated_at]
      properties:
        id: { type: integer }
        platform: { type: string }
        is_enabled: { type: boolean }
        post_time: { type: string }
        template: { type: string }
        account_handle: { type: string, nullable: true }
        updated_at: { type: string }
    SocialSettingsUpdate:
      type: object
      properties:
        is_enabled: { type: boolean }
        post_time: { type: string }
        template: { type: string }
        account_handle: { type: string }
...